LAPC-324 Add macros to enable and disable peripheral devices in devkit-e7.dts
Source: Alif Semiconductor
JIRA ID: LAPC-324
Type: Enhancement
Disposition: Local
Description:

The peripheral devices in the devkit-e7.dts file need to be
enabled/disabled based on the JSON file generated from the DCT tool.
To facilitate this each peripheral devices status is controlled
by the value of the macros, and the value of the macros is modified
based on the contents of the JSON file.

Reviewed-by: Hemanth Kumar <hemanth@alifsemi.com>
Signed-off-by: Jagadeesh Krishnanjanappa <jagadeesh@alifsemi.com>
Change-Id: I855a84cfe541e70eb2859a7c5e8bc86cb5abf2d5