LAPC-488 Remove hack in DW I3C driver
Source: Alif Semiconductor
JIRA ID: LAPC-488
Type: Development
Disposition: Local
Description:
Hack was added in dw-i3c-master.c to set the I2C clock speed to 2.5Khz.
This was added as a hack on an earlier A0 development board to get a
BNO055 chip working. The chip has been removed in the newer development
boards. So remove the hack as well.

Reviewed-by: Hemanth Kumar <hemanth@alifsemi.com>
Signed-off-by: Harith George <harith.g@alifsemi.com>
Change-Id: Iabad710d7293232d85bd40b254dfae01f3967980