LAPC-584 devkit-e7/appkit-e7:Use UART2 A as default serial console
Source: Alif Semiconductor
JIRA ID: LAPC-584
Type: Development
Disposition: Local
Description:
Use UART2 A as the default serial console for the APSS.

Set A32_UART to 4 in the Yocto config file to use UART4 B as the serial
console.
If building Linux as standalone, set UART4_STATUS to "okay" and
UART2_STATUS to "disabled" to use UART4 B as the serial console.

Reviewed-by: Hemanth Kumar <hemanth@alifsemi.com>
Reviewed-by: Harith G George <harith.g@alifsemi.com>
Signed-off-by: Jagadeesh Krishnanjanappa <jagadeesh@alifsemi.com>
Change-Id: Iaf3bcaaa196373534b64bcc6c1f9e93f33038524