PSBT-577 Fix CSI Camera initialization.
Source: Alif Semiconductor
JIRA ID: PSBT-577
Type: Defect Fix
Disposition: Local
Description:
CSI camera chain and hence the CPI initialization is failing as the
D-PHY is not achieving Stop state (LP11). This is happening as the PM
core is setting the ARX3A0 sensor to power-off state before the PHY init
of CSI happens which needs the DPHY lanes to be in LLP11 mode.

Fixing this by disabling the PM core on ARX3A0.

Reviewed-by: Harith George <harith.g@alifsemi.com>
Reviewed-by: Jagadeesh Krishnanjanappa <jagadeesh@alifsemi.com>
Reviewed-by: Hemanth Kumar <hemanth@alifsemi.com>
Signed-off-by: Yogender Kumar Arya <yogender.kumar@alifsemi.com>
Change-Id: I3f9fca2a34fbc393b2d059098d20f6e994fb0a9c