LAPC-660 Ensuring reliable I2C data transmission during probe
Source: Alif Semiconductor
JIRA ID: LAPC-660
Type: Defect Fix
Disposition: Local
Description:
During the probing of touchsensor GT911, a firmware image of 186 bytes
need to be sent across the I2C bus. This transmission happens during the
system init time as the touch sensor is compiled as a kernel component
and it happens from a separate workqueue thread, while the main probing
thread (swapper) is still probing other devices and doing some printks
on the minicom console. Due to the presence of printks from the swapper
thread, the servicing of I2C TX_BUFF_EMPTY interrupt, which refills the
I2C TX-FIFO with more contents. Due to this, the DW-I2C master generates
a STOP condition, as it expects that the transmission is complete and
there is no more data in tx_buff to transmit across the I2C line, but we
have still just sent around 32-bytes of data out of 186 bytes that
needed to be sent.

Hence, added an additional check in the ISR of I2C, which ensures that
the completion that the data-sending thread is holding is not released
in ISR just on the basis of STOP condition, but it also ensures that the
tx_buf that needs to be sent is also empty (transmission has completed).
This check will make sure that even if the I2C interrupts are delayed by
several milliseconds due to delay in printk statements, It will still be
serviced.

Reviewed-by: Harith George <harith.g@alifsemi.com>
Reviewed-by: Pankaj Pandey <pankaj.pandey@alifsemi.com>
Reviewed-by: Hemanth Kumar <hemanth@alifsemi.com>
Signed-off-by: Yogender Kumar Arya <yogender.kumar@alifsemi.com>