PSBT-680 Fix GT-911 DT node and reset sequence
Source: Alif Semiconductor
JIRA ID: PSBT-680
Type: Defect Fix
Disposition: Local
Description:
Fixing the DT node for GT911 sensor, so that irq-gpio is not marked as
IRQ_TYPE_EDGE_FALLING flag. This is because the irq-gpio is to be
programmed as output and later as input in the GT911 programming initial
reset sequence. When an IRQ_TYPE other than IRQ_TYPE_NONE is programmed
to the irq-gpio flags, the gpiod_direction_output call will fail as
GPIOs with IRQ flags can not change direction. This was creating
problems in the init sequence of the GT911 sensor.

Fixing the GT-911 reset sequence in goodix_reset. The sequence was not
as given in the data-sheet. With the updated sequence, the touch device
has been tested with both 0x14 and 0x5d slave-address.

Reviewed-by: Harith George <harith.g@alifsemi.com>
Reviewed-by: Hemanth Kumar <hemanth@alifsemi.com>
Signed-off-by: Yogender Kumar Arya <yogender.kumar@alifsemi.com>